Native Stork refactor
The current native scenario is too heavy in terms of memory and CPU usage.
This commit disables the current scenario for native mode and also create a new simpler scenario
in order to validate native execution
