+++ service-discovery/stork/src/test/java/io/quarkus/ts/stork/StorkLoadBalancerIT.java
@@ -5,44 +5,30 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.greaterThanOrEqualTo;
 import static org.hamcrest.Matchers.is;
 
-import java.io.IOException;
-import java.net.ServerSocket;
 import java.util.HashMap;
 import java.util.Map;
 
-import org.apache.http.HttpStatus;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 import io.quarkus.test.bootstrap.ConsulService;
 import io.quarkus.test.bootstrap.RestService;
 import io.quarkus.test.scenarios.QuarkusScenario;
+import io.quarkus.test.scenarios.annotations.DisabledOnNative;
 import io.quarkus.test.services.Container;
 import io.quarkus.test.services.QuarkusApplication;
 
-import junit.framework.AssertionFailedError;
-
 @QuarkusScenario
-public class StorkServiceDiscoveryIT {
+@DisabledOnNative(reason = "Too heavy native scenario in terms of memory consumption. Replaced by NativeStorkServiceDiscoveryIT")
+public class StorkLoadBalancerIT extends AbstractCommonTestCases {
 
     @Container(image = "${consul.image}", expectedLog = "Synced node info", port = 8500)
     static ConsulService consul = new ConsulService();
 
-    private static final String PREFIX = "ping-";
     private static final String DEFAULT_PONG_REPLICA_RESPONSE = "pongReplica";
     private static final String DEFAULT_PONG_RESPONSE = "pong";
-    private static final String PUNG_PORT = getAvailablePort();
     private static final String PONG_PORT = getAvailablePort();
     private static final String PONG_REPLICA_PORT = getAvailablePort();
-
-    @QuarkusApplication(classes = PungResource.class)
-    static RestService pungService = new RestService()
-            .withProperty("quarkus.http.port", PUNG_PORT)
-            .withProperty("pung-service-port", PUNG_PORT)
-            .withProperty("pung-service-host", "localhost")
-            .withProperty("stork.pung.service-discovery", "consul")
-            .withProperty("stork.pung.service-discovery.consul-port", () -> String.valueOf(consul.getPort()))
-            .withProperty("stork.pung.service-discovery.consul-host", () -> getConsultEndpoint(consul.getConsulEndpoint()));
 
     @QuarkusApplication(classes = PongResource.class)
     static RestService pongService = new RestService()
@@ -67,9 +53,6 @@ public class StorkServiceDiscoveryIT {
 
     @QuarkusApplication(classes = { PingResource.class, MyBackendPungProxy.class, MyBackendPongProxy.class })
     static RestService mainPingService = new RestService()
-            .withProperty("stork.pung.service-discovery", "consul")
-            .withProperty("stork.pung.service-discovery.consul-port", () -> String.valueOf(consul.getPort()))
-            .withProperty("stork.pung.service-discovery.consul-host", () -> getConsultEndpoint(consul.getConsulEndpoint()))
             .withProperty("stork.pong-replica.service-discovery", "consul")
             .withProperty("stork.pong-replica.service-discovery.refresh-period", "1")
             .withProperty("stork.pong-replica.load-balancer", "round-robin")
@@ -83,18 +66,12 @@ public class StorkServiceDiscoveryIT {
             .withProperty("stork.pong.service-discovery.consul-host", () -> getConsultEndpoint(consul.getConsulEndpoint()));
 
     @Test
-    public void invokeServiceByName() {
-        String response = makePingCall("pung");
-        assertThat("Service discovery by name fail.", PREFIX + "pung", is(response));
-    }
-
-    @Test
     public void storkLoadBalancerByRoundRobin() {
         Map<String, Integer> uniqueResp = new HashMap<>();
         final int requestAmount = 100;
         final int roundRobinError = (requestAmount / 2) - 1;
         for (int i = 0; i < requestAmount; i++) {
-            String response = makePingCall("pong");
+            String response = makePingCall(mainPingService, "pong").extract().body().asString();
             if (uniqueResp.containsKey(response)) {
                 uniqueResp.put(response, uniqueResp.get(response) + 1);
             } else {
@@ -109,25 +86,5 @@ public class StorkServiceDiscoveryIT {
                 is(greaterThanOrEqualTo(roundRobinError)));
         assertThat("Load balancer doesn't behaves as round-robin", uniqueResp.get(PREFIX + DEFAULT_PONG_REPLICA_RESPONSE),
                 is(greaterThanOrEqualTo(roundRobinError)));
-    }
-
-    private String makePingCall(String subPath) {
-        return mainPingService
-                .given()
-                .get("/ping/" + subPath).then()
-                .statusCode(HttpStatus.SC_OK)
-                .extract().body().asString();
-    }
-
-    private static String getConsultEndpoint(String endpoint) {
-        return endpoint.replaceFirst(":\\d+", "");
-    }
-
-    public static String getAvailablePort() {
-        try (ServerSocket socket = new ServerSocket(0)) {
-            return String.valueOf(socket.getLocalPort());
-        } catch (IOException e) {
-            throw new AssertionFailedError();
-        }
     }
 }
